Rename interface -> __public__
