+++ fiftyone/__init__.py
@@ -30,4 +30,4 @@ from pkgutil import extend_path
 #
 __path__ = extend_path(__path__, __name__)
 
-from fiftyone.interface import *
+from fiftyone.__public__ import *
